refactor(web): inline comment drawer content
